+++ pybamm/models/base_model.py
@@ -555,4 +555,3 @@ class BaseModel(object):
             return pybamm.IDAKLUSolver()
         else:
             return pybamm.CasadiSolver(mode="safe")
-